requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
accelerate>=0.24.0
sentencepiece>=0.1.99
pycocotools>=2.0.0
Pillow>=10.0.0